Reorganize some tests
